fix leftover issue from #247 if build_dir does not exist

+++ hyperglass/frontend/__init__.py
@@ -131,7 +131,10 @@ async def build_ui(app_path: Path):
         log.error(err)
         raise RuntimeError(str(err)) from err
 
-    shutil.rmtree(build_dir)
+    if build_dir.exists():
+        shutil.rmtree(build_dir)
+    else:
+        build_dir.mkdir()
     shutil.copytree(src=out_dir, dst=build_dir, dirs_exist_ok=False)
     log.bind(src=out_dir, dst=build_dir).debug("Migrated Next.JS build output")
